+++ extension/jsx/utils/effectsHelper.jsx
@@ -16,7 +16,9 @@ var bm_effectsHelper = (function () {
         layerIndex: 10,
         tint: 20,
         fill: 21,
-        stroke: 22
+        stroke: 22,
+        tritone: 23,
+        proLevels: 24
     };
     
     function getEffectType(name) {
@@ -27,6 +29,10 @@ var bm_effectsHelper = (function () {
             return effectTypes.fill;
         case 'ADBE Stroke':
             return effectTypes.stroke;
+        case 'ADBE Tritone':
+            return effectTypes.tritone;
+        case 'ADBE Pro Levels2':
+            return effectTypes.proLevels;
         default:
             return effectTypes.group;
         }
@@ -177,6 +183,7 @@ var bm_effectsHelper = (function () {
         ob.nm = elem.name;
         ob.mn = elem.matchName;
         ob.ix = elem.propertyIndex;
+        ob.en = elem.enabled === true ? 1 : 0;
         ob.ef = [];
         var i, len = elem.numProperties, prop;
         for (i = 0; i < len; i += 1) {
@@ -224,33 +231,13 @@ var bm_effectsHelper = (function () {
         for (i = 0; i < len; i += 1) {
             effectElement = effects(i + 1);
             var effectType = getEffectType(effectElement.matchName);
+            /*
             //If the effect is not a Slider Control and is not enabled, it won't be exported.
-            if(effectType !== effectTypes.group && !effects(i + 1).enabled){
-                //bm_eventDispatcher.log('PASO');
+            if(effectType !== effectTypes.group && !effectElement.enabled){
                 continue;
             }
+            */
             effectsArray.push(exportCustomEffect(effectElement ,effectType, frameRate));
-            /*var effectType = getEffectType(effectElement.matchName);
-            switch (effectType) {
-            case effectTypes.sliderControl:
-                effectsArray.push(exportSliderControl(effectElement.property('Slider'), frameRate));
-                break;
-            case effectTypes.angleControl:
-                effectsArray.push(exportAngleControl(effectElement.property('Angle'), frameRate));
-                break;
-            case effectTypes.colorControl:
-                effectsArray.push(exportColorControl(effectElement.property('Color'), frameRate));
-                break;
-            case effectTypes.pointControl:
-                effectsArray.push(exportPointControl(effectElement.property('Point'), frameRate));
-                break;
-            case effectTypes.checkboxControl:
-                effectsArray.push(exportCheckboxControl(effectElement.property('Checkbox'), frameRate));
-                break;
-            default:
-                //iterateEffectProperties(effectElement);
-                effectsArray.push(exportCustomEffect(effectElement, frameRate));
-            }*/
         }
         if (effectsArray.length) {
             layerData.ef = effectsArray;